Desafio 42 graphQL

+++ api/persistence/DAOs/products/mongoDbCloud.ts
@@ -44,7 +44,8 @@ class ProductMongoDAO extends IProductDAO {
 
   async addProduct(product: any): Promise<any | void> {
       const productToSave = new this.model(product)
-      await productToSave.save()
+      const data = await productToSave.save()
+      return new this.DTO(data).toJson()
   }
 
   public async updateProductById(id: any, newData: any): Promise<any> {
@@ -54,7 +55,7 @@ class ProductMongoDAO extends IProductDAO {
       if (updatedData.matchedCount === 0) {
           return { error: 'Product not found.' }
       } else {
-          return { msg: `Product ${id} updated!` }
+        return this.getProductById(id)
       }
      
   } catch (err) {
@@ -64,12 +65,13 @@ class ProductMongoDAO extends IProductDAO {
 
   public async deleteProductById(id: any): Promise<any> {
     try {
-      const deletedData = await this.model.deleteOne({ _id: id})
-      if (deletedData.deletedCount === 0) {
-        return { error: 'Product not found.' }
-    } else {
-        return { msg: 'Product deleted.' }
-    }
+      const entity = await this.getProductById(id)
+      
+      if (!entity) return undefined
+
+      await this.model.deleteOne({ _id: id })
+
+      return entity
   } catch (err) {
     Logger.error(`MongoAtlas deleteProductById method error: ${err}`) 
   }
